Conflicts:
	src/java/nxt/DbVersion.java

+++ src/java/nxt/BlockImpl.java
@@ -360,7 +360,7 @@ final class BlockImpl implements Block {
 
             BigInteger hit = new BigInteger(1, new byte[] {generationSignatureHash[7], generationSignatureHash[6], generationSignatureHash[5], generationSignatureHash[4], generationSignatureHash[3], generationSignatureHash[2], generationSignatureHash[1], generationSignatureHash[0]});
 
-            return Generator.verifyHit(hit, effectiveBalance, previousBlock, timestamp) || (this.height < Constants.TRANSPARENT_FORGING_BLOCK_5 && badBlocks.contains(this.getId()));
+            return Generator.verifyHit(hit, BigInteger.valueOf(effectiveBalance), previousBlock, timestamp) || (this.height < Constants.TRANSPARENT_FORGING_BLOCK_5 && badBlocks.contains(this.getId()));
 
         } catch (RuntimeException e) {
